requirements.txt
aiohttp
furl
loguru
PyYAML
qrcode
requests
rich
telebot
pyppeteer
pillow
opencv-python